Unit tests for config dump handler
