+++ test/server/admin/BUILD
@@ -42,10 +42,6 @@ envoy_cc_test(
         "//test/test_common:logging_lib",
         "//test/test_common:network_utility_lib",
         "//test/test_common:utility_lib",
-        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/filters/network/http_connection_manager/v3:pkg_cc_proto",
-        "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
     ],
 )
 
@@ -121,6 +117,14 @@ envoy_cc_test(
     deps = [
         ":admin_instance_lib",
         "@envoy_api//envoy/admin/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_test(
+    name = "config_dump_handler_test",
+    srcs = ["config_dump_handler_test.cc"],
+    deps = [
+        ":admin_instance_lib",
     ],
 )
